Moving IHttpClient
