--- AllReadyApp/Web-App/AllReady/Services/Routing/IHttpClient.cs
@@ -1,10 +1,0 @@
-﻿using System.Net.Http;
-using System.Threading.Tasks;
-
-namespace AllReady.Services.Routing
-{
-    public interface IHttpClient
-    {
-        Task<HttpResponseMessage> GetAsync(string requestUri);
-    }
-}
