--- AllReadyApp/Web-App/AllReady/Services/Routing/StaticHttpClient.cs
@@ -1,15 +1,0 @@
-﻿using System.Net.Http;
-using System.Threading.Tasks;
-
-namespace AllReady.Services.Routing
-{
-    public class StaticHttpClient : IHttpClient
-    {
-        private static readonly HttpClient HttpClient = new HttpClient();
-
-        public async Task<HttpResponseMessage> GetAsync(string requestUri)
-        {
-            return await HttpClient.GetAsync(requestUri);
-        }
-    }
-}
